test effectuer pour ActivityController/UserController/EtablissementController, changepassword de userController pas tester encore

+++ src/main/java/org/rina/controller/UserController.java
@@ -1,7 +1,10 @@
 package org.rina.controller;
 
 import org.rina.config.JwtService;
+import org.rina.dto.request.UserDto;
+
 import jakarta.servlet.http.HttpServletRequest;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.rina.model.User;
 import org.rina.service.UserService;
@@ -18,37 +21,68 @@ import java.util.Optional;
 @RestController
 @RequestMapping("/api/v1/users")
 @RequiredArgsConstructor
+@SuppressWarnings("unused")
 public class UserController {
 
     private final UserService userService;
     private final JwtService jwtService;
     private final PasswordEncoder passwordEncoder;
 
+//    @GetMapping
+//    public ResponseEntity<List<User>> getAll() {
+//        List<User> users = userService.findAll();
+//        return ResponseEntity.ok(users);
+//    }
+    
     @PostMapping
-    public ResponseEntity<User> create(@RequestBody User request) {
-        User user = User.builder()
-                .username(request.getUsername())
-                .password(passwordEncoder.encode(request.getPassword()))
-                .role(request.getRole())
-                .build();
-        User savedUser = userService.save(user);
-
-        return ResponseEntity.ok(savedUser);
+    public ResponseEntity<String> createUser(@Valid @RequestBody UserDto userDto) {
+    	// Vérifie qu'il n'existe pas déjà cet username
+    	Optional<User> exitingUser = userService.findByUsername(userDto.getUsername());
+    	if (exitingUser.isEmpty()) {
+    		
+    		userDto.setPassword(passwordEncoder.encode(userDto.getPassword()));
+    		User newUser = userService.insert(userDto.toUser());
+    		return ResponseEntity.ok().build();
+        } 
+        
+        else return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("L'utilisateur existe déjà.");
     }
+//    	User user = User.builder()
+//                .username(userDto.getUsername())
+//                .password(passwordEncoder.encode(userDto.getPassword()))
+//                .role(userDto.getRole())
+//                .build();
+//        User savedUser = userService.save(user);
+//
+//        return ResponseEntity.ok(savedUser);
+    
 
     @PutMapping("/{username}")
-    public ResponseEntity<User> update(@PathVariable String username, @RequestBody User request) {
-        User userUpdate = userService.findByUsername(username).get();
-        userUpdate.setUsername(request.getUsername());
-        userUpdate.setPassword(passwordEncoder.encode(request.getPassword()));
-        userUpdate.setRole(request.getRole());
-        User savedUser = userService.save(userUpdate);
-        return ResponseEntity.ok(savedUser);
-    }
+    public ResponseEntity<User> newPassword(@PathVariable String username, @Valid @RequestBody UserDto userDto) {
+    	// Vérifie d'abord si l'utilisateur existe en fonction du username
+    	Optional<User> existingUser = userService.findByUsername(username);
+    	
+    	if (existingUser.isPresent()) {
+    		
+    		// Mise à jour l'utilisateur existant avec les nouvelles valeurs
+        	userDto.setUsername(username);
+        	userDto.setPassword(passwordEncoder.encode(userDto.getPassword()));
+        	User updateUser = userService.updateUser(username, userDto.toUser());
+            return ResponseEntity.ok().build();
+    	}
+    
+    	else return ResponseEntity.notFound().build();
+    }       
+    //userUpdate.setUsername(userDto.getUsername());
+     //userUpdate.setPassword(passwordEncoder.encode(userDto.getPassword()));
+     // userUpdate.setRole(userDto.getRole());
+        //User savedUser = userService.save(userDto.toUser());
+        //return ResponseEntity.ok(savedUser);
+   
 
-    @GetMapping("/username")
-    public ResponseEntity<User> getByUsername(HttpServletRequest request) {
-        String token = request.getHeader(HttpHeaders.AUTHORIZATION);
+    @GetMapping
+    public ResponseEntity<User> getUsername(HttpServletRequest userDto) {
+        String token = userDto.getHeader(HttpHeaders.AUTHORIZATION);
         if(token==null || !token.startsWith("Bearer ")){
             return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
@@ -62,20 +96,18 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<User> getById(@PathVariable Long id) {
+    public ResponseEntity<User> getUserById(@PathVariable Long id) {
         Optional<User> user = userService.findById(id);
         return ResponseEntity.of(user);
     }
 
-    @GetMapping
-    public ResponseEntity<List<User>> getAll() {
-        List<User> users = userService.findAll();
-        return ResponseEntity.ok(users);
-    }
-
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable Long id) {
-        userService.deleteById(id);
-        return ResponseEntity.ok().build();
-    }
+	    if (userService.existsById(id)) {
+	    	userService.deleteById(id);
+			return ResponseEntity.ok().build();
+		}
+		
+		 else return ResponseEntity.notFound().build();
+	}
 }
